feat(nargo): perform initial split into core + cli
