+++ crates/nargo/src/cli/execute_cmd.rs
@@ -67,7 +67,7 @@ pub(crate) fn execute_program(
 ) -> Result<WitnessMap, CliError> {
     let mut solved_witness = compiled_program.abi.encode(inputs_map, None)?;
 
-    let backend = crate::backends::ConcreteBackend;
+    let backend = nargo_core::backends::ConcreteBackend;
     let mut blocks = Blocks::default();
     let (unresolved_opcodes, oracles) = backend.solve(
         &mut solved_witness,
