+++ crates/nargo_core/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "nargo_core"
+description = "Noir's package manager"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[build-dependencies]
+rustc_version = "0.4.0"
+
+[dependencies]
+cfg-if.workspace = true
+toml.workspace = true
+serde.workspace = true
+thiserror.workspace = true
+
+# Backends
+aztec_backend = { optional = true, package = "barretenberg_static_lib", git = "https://github.com/noir-lang/aztec_backend", rev = "26178359a2251e885f15f0a4d1a686afda04aec9" }
+aztec_wasm_backend = { optional = true, package = "barretenberg_wasm", git = "https://github.com/noir-lang/aztec_backend", rev = "26178359a2251e885f15f0a4d1a686afda04aec9" }
+marlin_arkworks_backend = { optional = true, git = "https://github.com/noir-lang/marlin_arkworks_backend", rev = "144378edad821bfaa52bf2cacca8ecc87514a4fc" }
+
+[features]
+default = ["plonk_bn254"]
+# The plonk backend can only use bn254, so we do not specify the field
+plonk_bn254 = ["aztec_backend"]
+marlin = ["marlin_arkworks_backend/bls12_381"]
+plonk_bn254_wasm = ["aztec_wasm_backend"]
