+++ crates/nargo_core/src/lib.rs
@@ -1,0 +1,10 @@
+#![forbid(unsafe_code)]
+#![warn(unused_crate_dependencies, unused_extern_crates)]
+#![warn(unreachable_pub)]
+
+//! Nargo is the package manager for Noir
+//! This name was used because it sounds like `cargo` and
+//! Noir Package Manager abbreviated is npm, which is already taken.
+
+pub mod backends;
+pub mod manifest;
