+++ crates/nargo_core/src/workspace.rs
